+++ tests/index.test.ts
@@ -1,11 +1,6 @@
 import { describe, it, expect, beforeAll } from "vitest";
 
 describe("pixel script", () => {
-	beforeAll(async () => {
-		// Vitest will have jsdom ready here
-		await import("../src/index");
-	});
-
 	it("should pass this dummy test", () => {
 		expect(true).toBe(true);
 	});
