+++ vitest.config.ts
@@ -1,6 +1,8 @@
 import { defineConfig } from "vitest/config";
+import tsconfigPaths from "vite-tsconfig-paths";
 
 export default defineConfig({
+	plugins: [tsconfigPaths()],
 	test: {
 		include: ["tests/**/*.test.ts"],
 		globals: true,
